# api/openapi.yml
openapi: 3.0.3
info:
  title: SHINE Movie Database API
  description: |
    SHINE is a comprehensive movie database API designed to be the world's most organized movie database. 
    It collects and organizes movie information, awards, nominations, and multilingual translations.

    **Features:**
    - Date-seeded movie selections (daily/weekly/monthly)
    - Comprehensive awards and nominations tracking
    - Multilingual movie title support
    - User-submitted article links
    - Admin management interface

    **Authentication:**
    - Public endpoints require no authentication
    - Admin endpoints require JWT Bearer token

    **Rate Limiting:**
    - Article link submissions: 10 per hour per IP
  version: 1.0.0
  contact:
    name: SHINE Movie Database
    url: https://shine-film.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.shine-film.com
    description: Production server
paths:
  /:
    get:
      summary: Get date-seeded movie selections
      description: Returns daily, weekly, and monthly movie selections using deterministic hash-based algorithms
      operationId: getMovieSelections
      tags:
        - Movies
      security: []
      parameters:
        - name: locale
          in: query
          description: Language preference for movie titles
          required: false
          schema:
            type: string
            enum: [en, ja]
            default: en
      responses:
        '200':
          description: Successfully retrieved movie selections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieSelections'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /movies/search:
    get:
      summary: Search movies
      description: Search and filter movies with pagination support
      operationId: searchMovies
      tags:
        - Movies
      security: []
      parameters:
        - name: q
          in: query
          description: Search query for movie titles
          required: false
          schema:
            type: string
            maxLength: 100
        - name: year
          in: query
          description: Filter by release year
          required: false
          schema:
            type: integer
            minimum: 1900
            maximum: 2100
        - name: language
          in: query
          description: Filter by original language
          required: false
          schema:
            type: string
            pattern: '^[a-z]{2}$'
        - name: hasAwards
          in: query
          description: Filter movies with awards/nominations
          required: false
          schema:
            type: boolean
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Successfully retrieved search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieSearchResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /movies/{id}:
    get:
      summary: Get movie details
      description: Retrieve detailed information about a specific movie including translations, posters, and nominations
      operationId: getMovieById
      tags:
        - Movies
      security: []
      parameters:
        - name: id
          in: path
          description: Movie unique identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved movie details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetails'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /auth/login:
    post:
      summary: Admin login
      description: Authenticate as admin and receive JWT token
      operationId: adminLogin
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  description: Admin password
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token (valid for 24 hours)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
    Movie:
      type: object
      properties:
        uid:
          type: string
          format: uuid
          description: Unique identifier for the movie
        year:
          type: integer
          minimum: 1900
          maximum: 2100
          description: Release year
        originalLanguage:
          type: string
          pattern: '^[a-z]{2}$'
          description: ISO 639-1 language code
        imdbId:
          type: string
          pattern: '^tt\d{7,8}$'
          nullable: true
          description: IMDb identifier (tt format)
        tmdbId:
          type: integer
          minimum: 1
          nullable: true
          description: TMDb identifier
        title:
          type: string
          description: Movie title (localized)
        posterUrl:
          type: string
          format: uri
          nullable: true
          description: Primary poster image URL
        imdbUrl:
          type: string
          format: uri
          nullable: true
          description: IMDb page URL
        hasAwards:
          type: boolean
          description: Whether the movie has awards or nominations
      required:
        - uid
        - year
        - originalLanguage
        - title
    MovieDetails:
      allOf:
        - $ref: '#/components/schemas/Movie'
        - type: object
          properties:
            posters:
              type: array
              items:
                $ref: '#/components/schemas/Poster'
              description: All poster images for the movie
            translations:
              type: array
              items:
                $ref: '#/components/schemas/Translation'
              description: Movie title translations
            nominations:
              type: array
              items:
                $ref: '#/components/schemas/Nomination'
              description: Awards and nominations
    Poster:
      type: object
      properties:
        uid:
          type: string
          format: uuid
          description: Unique identifier for the poster
        url:
          type: string
          format: uri
          description: Poster image URL
        width:
          type: integer
          minimum: 1
          nullable: true
          description: Image width in pixels
        height:
          type: integer
          minimum: 1
          nullable: true
          description: Image height in pixels
        languageCode:
          type: string
          pattern: '^[a-z]{2}$'
          nullable: true
          description: Language-specific poster
        sourceType:
          type: string
          enum: [tmdb, manual, scraper]
          description: Source of the poster
        isPrimary:
          type: boolean
          description: Whether this is the primary poster
      required:
        - uid
        - url
        - sourceType
        - isPrimary
    Translation:
      type: object
      properties:
        languageCode:
          type: string
          pattern: '^[a-z]{2}$'
          description: ISO 639-1 language code
        content:
          type: string
          description: Translated movie title
        isDefault:
          type: boolean
          description: Whether this is the default translation
      required:
        - languageCode
        - content
        - isDefault
    Nomination:
      type: object
      properties:
        uid:
          type: string
          format: uuid
          description: Unique identifier for the nomination
        isWinner:
          type: boolean
          description: Whether this nomination won the award
        specialMention:
          type: string
          nullable: true
          description: Special mention or note about the nomination
        category:
          $ref: '#/components/schemas/AwardCategory'
        ceremony:
          $ref: '#/components/schemas/AwardCeremony'
        organization:
          $ref: '#/components/schemas/AwardOrganization'
      required:
        - uid
        - isWinner
        - category
        - ceremony
        - organization
    AwardCategory:
      type: object
      properties:
        uid:
          type: string
          format: uuid
          description: Unique identifier for the category
        name:
          type: string
          description: Category name
      required:
        - uid
        - name
    AwardCeremony:
      type: object
      properties:
        uid:
          type: string
          format: uuid
          description: Unique identifier for the ceremony
        number:
          type: integer
          minimum: 1
          description: Ceremony number (e.g., 95th Academy Awards)
        year:
          type: integer
          minimum: 1900
          maximum: 2100
          description: Ceremony year
      required:
        - uid
        - number
        - year
    AwardOrganization:
      type: object
      properties:
        uid:
          type: string
          format: uuid
          description: Unique identifier for the organization
        name:
          type: string
          description: Organization full name
        shortName:
          type: string
          nullable: true
          description: Organization short name or abbreviation
      required:
        - uid
        - name
    ArticleLink:
      type: object
      properties:
        uid:
          type: string
          format: uuid
          description: Unique identifier for the article link
        url:
          type: string
          format: uri
          description: Article URL
        title:
          type: string
          description: Article title
        description:
          type: string
          nullable: true
          description: Article description
        submittedAt:
          type: string
          format: date-time
          description: Submission timestamp
      required:
        - uid
        - url
        - title
        - submittedAt
    MovieSelections:
      type: object
      properties:
        daily:
          $ref: '#/components/schemas/MovieWithDetails'
        weekly:
          $ref: '#/components/schemas/MovieWithDetails'
        monthly:
          $ref: '#/components/schemas/MovieWithDetails'
      required:
        - daily
        - weekly
        - monthly
    MovieWithDetails:
      allOf:
        - $ref: '#/components/schemas/Movie'
        - type: object
          properties:
            nominations:
              type: array
              items:
                $ref: '#/components/schemas/Nomination'
              description: Awards and nominations
            articleLinks:
              type: array
              items:
                $ref: '#/components/schemas/ArticleLink'
              description: Related article links (max 3)
    MovieSearchResponse:
      type: object
      properties:
        movies:
          type: array
          items:
            $ref: '#/components/schemas/Movie'
          description: Search results
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
        filters:
          $ref: '#/components/schemas/SearchFilters'
      required:
        - movies
        - pagination
        - filters
...
    PaginationInfo:
      type: object
      properties:
        currentPage:
          type: integer
          minimum: 1
          description: Current page number
        totalPages:
          type: integer
          minimum: 1
          description: Total number of pages
        totalCount:
          type: integer
          minimum: 0
          description: Total number of items
        hasNextPage:
          type: boolean
          description: Whether there is a next page
        hasPrevPage:
          type: boolean
          description: Whether there is a previous page
      required:
        - currentPage
        - totalPages
        - totalCount
        - hasNextPage
        - hasPrevPage
...
    SearchFilters:
      type: object
      properties:
        query:
          type: string
          nullable: true
          description: Applied search query
        year:
          type: integer
          nullable: true
          description: Applied year filter
        language:
          type: string
          nullable: true
          description: Applied language filter
        hasAwards:
          type: boolean
          nullable: true
          description: Applied awards filter
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          nullable: true
          description: Additional error details
      required:
        - error
        - code
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'Validation failed'
            code: 'VALIDATION_ERROR'
            details:
              field: 'Field validation message'
    UnauthorizedError:
      description: Authentication error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'Authentication required'
            code: 'AUTHENTICATION_ERROR'
            details:
              reason: 'MISSING_TOKEN'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'Resource not found'
            code: 'NOT_FOUND'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'Internal server error'
            code: 'INTERNAL_ERROR'
...